vnf_detector/requirements.txt
APScheduler==3.5.3
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==1.11.20
djangorestframework==3.9.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
http==0.2
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyaml==18.11.0
Pygments==2.3.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.24.1
wcwidth==0.1.7
